revert annotationprocessor to jdk 8
