+++ annotationprocessor/build.gradle
@@ -26,5 +26,5 @@ dependencies {
     compile project(':acra-javacore')
 }
 
-sourceCompatibility = "1.9"
-targetCompatibility = "1.9"
+sourceCompatibility = "1.8"
+targetCompatibility = "1.8"
