+++ gradle.properties
@@ -18,7 +18,7 @@
 
 # gradle options
 org.gradle.daemon=true
-org.gradle.jvmargs=-Xmx4g -agentlib:jdwp=transport=dt_socket,server=y,suspend=n,address=5005 --add-modules java.se.ee --add-opens java.base/jdk.internal.misc=ALL-UNNAMED
+org.gradle.jvmargs=-Xmx4g -agentlib:jdwp=transport=dt_socket,server=y,suspend=n,address=5005
 org.gradle.configureondemand=true
 org.gradle.parallel=true
 org.gradle.caching=true
@@ -30,7 +30,7 @@ version=5.1.0-SNAPSHOT
 supportVersion=27.0.2
 androidVersion=27
 androidMinVersion=14
-androidBuildPluginVersion=3.1.0-beta1
+androidBuildPluginVersion=3.2.0-alpha01
 bintrayPluginVersion=1.8.0
 releasePluginVersion=2.6.0
 autoServiceVersion=1.0-rc4
